utils/resume_parser.py
import PyPDF2
import docx

def parse_resume(file):
    filename = file.filename.lower()
    
    if filename.endswith('.pdf'):
        return parse_pdf(file)
    elif filename.endswith('.docx'):
        return parse_docx(file)
    else:
        raise ValueError(f"Unsupported file format: {filename.split('.')[-1]}. Please upload a PDF or DOCX file.")

def parse_pdf(file):
    try:
        pdf_reader = PyPDF2.PdfReader(file)
        text = ""
        for page in pdf_reader.pages:
            text += page.extract_text()
        return text
    except Exception as e:
        raise ValueError(f"Error parsing PDF file: {str(e)}")

def parse_docx(file):
    try:
        doc = docx.Document(file)
        text = ""
        for paragraph in doc.paragraphs:
            text += paragraph.text + "\n"
        return text
    except Exception as e:
        raise ValueError(f"Error parsing DOCX file: {str(e)}")